feat: :art: improve visibility of index page
divide button from index to EditProfile, add EditProfileModal
