+++ frontend/src/pages/ProfilePage/index.tsx
@@ -1,54 +1,34 @@
-import { useState } from 'react';
 import { useParams } from 'react-router-dom';
 
-import { Modal } from 'common';
 import { useGetUser } from 'hooks';
 
 import { ProfileCard } from './ProfileCard';
-import { EditProfile } from './EditProfile';
 import { ManageFriends } from './ManageFriends';
 import { MatchHistoryList } from './MatchHistoryList';
 import { AchievementList } from './AchievementList';
 import { TwoFactorAuth } from './TwoFactorAuth';
+import { EditProfile } from './EditProfile';
 
 export const ProfilePage = () => {
-  const [isModalShown, setModalShown] = useState<Boolean>(false);
   const { id } = useParams();
   const { data: profile } = useGetUser(id);
   const { data: myProfile } = useGetUser();
 
-  function handleOpenModal() {
-    setModalShown(true);
-  }
-
-  function handleCloseModal() {
-    setModalShown(false);
-  }
-
   if (!profile) return <span>error</span>;
   return (
-    <>
-      <main>
-        <h1>Profile Page</h1>
-        <ProfileCard user={profile} />
-        {!id || profile.id === myProfile.id ? (
-          <>
-            <button type="button" onClick={handleOpenModal}>
-              Edit Profile
-            </button>
-            <TwoFactorAuth />
-          </>
-        ) : (
-          <ManageFriends user={profile} />
-        )}
-        <MatchHistoryList userId={profile.id} />
-        <AchievementList userId={profile.id} />
-      </main>
-      {isModalShown && (
-        <Modal onClickClose={handleCloseModal}>
-          <EditProfile onClickClose={handleCloseModal} />
-        </Modal>
+    <main>
+      <h1>Profile Page</h1>
+      <ProfileCard user={profile} />
+      {!id || profile.id === myProfile.id ? (
+        <>
+          <EditProfile />
+          <TwoFactorAuth />
+        </>
+      ) : (
+        <ManageFriends user={profile} />
       )}
-    </>
+      <MatchHistoryList userId={profile.id} />
+      <AchievementList userId={profile.id} />
+    </main>
   );
 };
